Refactor to EmulatorInfoBar component
