+++ src/scripts/emulator-view.ts
@@ -10,9 +10,7 @@ export class EmulatorView {
   screen: JQuery<HTMLElement> // screen element
   testcard: JQuery<HTMLElement> // testcard element
   canvas: EmulatorCanvas
-  emuStatus: HTMLElement // = document.getElementById('emu_status');
-  showCoords: boolean
-  emulator: Emulator | undefined
+  emulator: Emulator | undefined // Dont hold references to the emulator, it may be paused and destroyed
   model: Model | undefined
 
   constructor() {
@@ -29,23 +27,11 @@ export class EmulatorView {
     //this.screen = document.getElementById('screen');
     //		const screen = document.getElementById('screen'); // this.root.find('.screen');
     this.canvas = bestCanvas(screen[0])
-    const emuStatus = document.getElementById('emu_status')
-    if (!emuStatus) {
-      throw new Error('No emu_status element found')
-    }
-    this.emuStatus = emuStatus
-    this.showCoords = false // coordinate display mode
-
-    // coords handlers
-    screen.mousemove((event: any) => this.mouseMove(event))
-    screen.mouseleave(() => this.mouseLeave())
 
     // forward key events to emulator
     screen.keyup((event: any) => this.emulator?.keyUp(event))
     screen.keydown((event: any) => this.emulator?.keyDown(event))
     screen.blur(() => this.emulator?.clearKeys())
-
-    setInterval(this.timer.bind(this), 1000)
   }
 
   async boot(model: Model) {
@@ -87,78 +73,5 @@ export class EmulatorView {
       this.screen.show()
       this.testcard.hide()
     }
-  }
-
-  private timer() {
-    if (this.emulator && !this.showCoords) {
-      this.emuStatus.innerHTML = `${this.model?.name} | ${Math.floor(this.emulator.cpu.currentCycles / 2000000)} s`
-    }
-  }
-
-  private mouseLeave() {
-    this.showCoords = false
-    this.timer()
-  }
-
-  private mouseMove(event: any) {
-    if (!this.emulator) return
-    this.showCoords = true
-    const processor = this.emulator.cpu
-    const screen = this.screen // this.root.find('.screen');
-    const screenMode = processor.readmem(0x0355)
-    let W
-    let H
-    let graphicsMode = true
-    switch (screenMode) {
-      case 0:
-        W = 80
-        H = 32
-        break
-      case 1:
-      case 4:
-        W = 40
-        H = 32
-        break
-      case 2:
-      case 5:
-        W = 20
-        H = 32
-        break
-      case 3:
-        W = 80
-        H = 25.6
-        graphicsMode = false
-        break
-      case 6:
-        W = 40
-        H = 25.6
-        graphicsMode = false
-        break
-      case 7:
-        W = 40
-        H = 25.6
-        graphicsMode = false
-        break
-      default:
-        // Unknown screen mode!
-        return
-    }
-    // 8 and 16 here are fudges to allow for a margin around the screen
-    // canvas - not sure exactly where that comes from...
-    let x = event.offsetX - 8
-    let y = event.offsetY - 8
-    const sw = (screen.width() ?? 16) - 16
-    const sh = (screen.height() ?? 16) - 16
-    const X = Math.floor((x * W) / sw)
-    const Y = Math.floor((y * H) / sh)
-    let html = `Text: (${X},${Y})`
-    if (graphicsMode) {
-      // Graphics Y increases up the screen.
-      y = sh - y
-      x = Math.floor((x * 1280) / sw)
-      y = Math.floor((y * 1024) / sh)
-      html += ` &nbsp; Graphics: (${x},${y})`
-    }
-    this.emuStatus.innerHTML = html
   }
 }
